services/audio-renderer-service/pyproject.toml
[project]
name = "audio-renderer-service"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "motor>=3.6",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "azure-cognitiveservices-speech>=1.43",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "httpx>=0.27",
    "mongomock-motor>=0.0.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
exclude = ["**/__pycache__"]
typeCheckingMode = "standard"
reportMissingTypeStubs = false
pythonVersion = "3.13"
